requirements.txt
kafka-python==2.0.2
opensearch-py==2.3.1
faker==19.6.2
requests==2.31.0
pandas==2.1.1
streamlit==1.26.0
plotly==5.16.1
scikit-learn==1.3.0
